Move upsert/delete purpose functions to `purpose-rmw-sql`
