+++ packages/datalake-data-export/test/utils.ts
@@ -15,11 +15,13 @@ import {
 } from "pagopa-interop-commons";
 import {
   catalogReadModelServiceBuilder,
-  purposeReadModelServiceBuilder,
   tenantReadModelServiceBuilder,
 } from "pagopa-interop-readmodel";
 import { Agreement, EService, Purpose, Tenant } from "pagopa-interop-models";
-import { upsertAgreement } from "pagopa-interop-readmodel/testUtils";
+import {
+  upsertAgreement,
+  upsertPurpose,
+} from "pagopa-interop-readmodel/testUtils";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { readModelServiceBuilderSQL } from "../src/services/readModelServiceSQL.js";
 import { config } from "../src/config/config.js";
@@ -46,7 +48,6 @@ export const purposes: PurposeCollection = readModelRepository.purposes;
 const oldReadModelService = readModelServiceBuilder(readModelRepository);
 
 const catalogReadModelServiceSQL = catalogReadModelServiceBuilder(readModelDB);
-const purposeReadModelServiceSQL = purposeReadModelServiceBuilder(readModelDB);
 const tenantReadModelServiceSQL = tenantReadModelServiceBuilder(readModelDB);
 const readModelServiceSQL = readModelServiceBuilderSQL(readModelDB);
 export const readModelService =
@@ -80,7 +81,7 @@ export const seedAgreements = async (
 
 export const seedPurposes = async (purposes: Purpose[]): Promise<void> => {
   for (const p of purposes) {
-    await purposeReadModelServiceSQL.upsertPurpose(p, 0);
+    await upsertPurpose(readModelDB, p, 0);
   }
 };
