+++ packages/purpose-readmodel-writer-sql/src/purposeConsumerServiceV2.ts
@@ -5,11 +5,11 @@ import {
   unsafeBrandId,
 } from "pagopa-interop-models";
 import { match } from "ts-pattern";
-import { CustomReadModelService } from "./readModelService.js";
+import { PurposeWriterService } from "./purposeWriterService.js";
 
 export async function handleMessageV2(
   message: PurposeEventEnvelopeV2,
-  purposeReadModelService: CustomReadModelService
+  purposeWriterService: PurposeWriterService
 ): Promise<void> {
   const purposeV2 = message.data.purpose;
   if (!purposeV2) {
@@ -23,7 +23,7 @@ export async function handleMessageV2(
       { type: "WaitingForApprovalPurposeDeleted" },
       { type: "PurposeDeletedByRevokedDelegation" },
       async (message) => {
-        await purposeReadModelService.deletePurposeById(
+        await purposeWriterService.deletePurposeById(
           unsafeBrandId(message.stream_id),
           message.version
         );
@@ -48,7 +48,7 @@ export async function handleMessageV2(
       { type: "PurposeCloned" },
       { type: "PurposeVersionArchivedByRevokedDelegation" },
       async (message) => {
-        await purposeReadModelService.upsertPurpose(purpose, message.version);
+        await purposeWriterService.upsertPurpose(purpose, message.version);
       }
     )
     .exhaustive();
