+++ packages/readmodel/src/purposeReadModelService.ts
@@ -1,4 +1,4 @@
-import { and, eq, lte, SQL } from "drizzle-orm";
+import { and, eq, SQL } from "drizzle-orm";
 import {
   genericInternalError,
   Purpose,
@@ -13,73 +13,16 @@ import {
   purposeVersionDocumentInReadmodelPurpose,
   purposeVersionInReadmodelPurpose,
 } from "pagopa-interop-readmodel-models";
-import { splitPurposeIntoObjectsSQL } from "./purpose/splitters.js";
 import {
   aggregatePurpose,
   aggregatePurposeArray,
   toPurposeAggregator,
   toPurposeAggregatorArray,
 } from "./purpose/aggregators.js";
-import { checkMetadataVersion } from "./utils.js";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function purposeReadModelServiceBuilder(db: DrizzleReturnType) {
   return {
-    async upsertPurpose(
-      purpose: Purpose,
-      metadataVersion: number
-    ): Promise<void> {
-      await db.transaction(async (tx) => {
-        const shouldUpsert = await checkMetadataVersion(
-          tx,
-          purposeInReadmodelPurpose,
-          metadataVersion,
-          purpose.id
-        );
-
-        if (!shouldUpsert) {
-          return;
-        }
-
-        await tx
-          .delete(purposeInReadmodelPurpose)
-          .where(eq(purposeInReadmodelPurpose.id, purpose.id));
-
-        const {
-          purposeSQL,
-          riskAnalysisFormSQL,
-          riskAnalysisAnswersSQL,
-          versionsSQL,
-          versionDocumentsSQL,
-        } = splitPurposeIntoObjectsSQL(purpose, metadataVersion);
-
-        await tx.insert(purposeInReadmodelPurpose).values(purposeSQL);
-
-        if (riskAnalysisFormSQL) {
-          await tx
-            .insert(purposeRiskAnalysisFormInReadmodelPurpose)
-            .values(riskAnalysisFormSQL);
-        }
-
-        if (riskAnalysisAnswersSQL) {
-          for (const riskAnalysisAnswerSQL of riskAnalysisAnswersSQL) {
-            await tx
-              .insert(purposeRiskAnalysisAnswerInReadmodelPurpose)
-              .values(riskAnalysisAnswerSQL);
-          }
-        }
-
-        for (const versionSQL of versionsSQL) {
-          await tx.insert(purposeVersionInReadmodelPurpose).values(versionSQL);
-        }
-
-        for (const versionDocumentSQL of versionDocumentsSQL) {
-          await tx
-            .insert(purposeVersionDocumentInReadmodelPurpose)
-            .values(versionDocumentSQL);
-        }
-      });
-    },
     async getPurposeById(
       purposeId: PurposeId
     ): Promise<WithMetadata<Purpose> | undefined> {
@@ -209,22 +152,8 @@ export function purposeReadModelServiceBuilder(db: DrizzleReturnType) {
 
       return aggregatePurposeArray(toPurposeAggregatorArray(queryResult));
     },
-    async deletePurposeById(
-      purposeId: PurposeId,
-      version: number
-    ): Promise<void> {
-      await db
-        .delete(purposeInReadmodelPurpose)
-        .where(
-          and(
-            eq(purposeInReadmodelPurpose.id, purposeId),
-            lte(purposeInReadmodelPurpose.metadataVersion, version)
-          )
-        );
-    },
   };
 }
-
 export type PurposeReadModelService = ReturnType<
   typeof purposeReadModelServiceBuilder
 >;
